+++ odfdo/document.py
@@ -30,6 +30,7 @@ import posixpath
 from contextlib import suppress
 from copy import deepcopy
 from importlib import resources as rso
+from itertools import chain
 from mimetypes import guess_type
 from operator import itemgetter
 from pathlib import Path
@@ -547,7 +548,28 @@ class Document:
         # And the description
         print_info("Description", meta.get_description())
 
-        return "\n".join(result) + "\n"
+        return "\n".join(result)
+
+    def to_markdown(self) -> str:
+        doc_type = self.get_type()
+        if doc_type not in {
+            "text",
+        }:
+            raise NotImplementedError(
+                f"Type of document '{doc_type}' not supported yet"
+            )
+        md_list = self._document_to_markdown()
+        raw_text = "\n".join(x for x in md_list)
+        return "\n".join(x.rstrip(" ") for x in raw_text.split("\n"))
+
+    def _document_to_markdown(self) -> list[str]:
+        return [
+            item
+            for item in chain.from_iterable(
+                child._to_markdown() for child in self.body.children
+            )
+            if item
+        ]
 
     def add_file(self, path_or_file: str | Path) -> str:
         """Insert a file from a path or a file-like object in the container.